Rename assembly and namespace to SemanticVersioning to match nuget package
