+++ src/SemanticVersioning/Desugarer.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace SemVer
+namespace SemanticVersioning
 {
     internal static class Desugarer
     {
