+++ src/SemanticVersioning/Properties/AssemblyInfo.cs
@@ -7,9 +7,9 @@ using System.Runtime.InteropServices;
 // associated with an assembly.
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("SemVer")]
+[assembly: AssemblyProduct("SemanticVersioning")]
 [assembly: AssemblyTrademark("")]
-[assembly: InternalsVisibleTo("SemVer.Tests")]
+[assembly: InternalsVisibleTo("SemanticVersioning.Tests")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
 // to COM components.  If you need to access a type in this assembly from
