+++ src/SemanticVersioning/Range.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace SemVer
+namespace SemanticVersioning
 {
     /// <summary>
     /// Specifies valid versions.
