+++ src/SemanticVersioning/SemanticVersioning.csproj
@@ -14,7 +14,7 @@
     <AssemblyTitle>SemanticVersioning</AssemblyTitle>
     <VersionPrefix>1.0.0</VersionPrefix>
     <Authors>Adam Reeve</Authors>
-    <AssemblyName>SemVer</AssemblyName>
+    <AssemblyName>SemanticVersioning</AssemblyName>
     <PackageId>SemanticVersioning</PackageId>
     <PackageProjectUrl>https://github.com/adamreeve/semver.net</PackageProjectUrl>
     <PackageLicenseUrl>https://github.com/adamreeve/semver.net/blob/master/LICENSE</PackageLicenseUrl>
