+++ test/SemanticVersioning.Tests/AdvancedRanges.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using Xunit;
 using Xunit.Extensions;
 
-namespace SemVer.Tests
+namespace SemanticVersioning.Tests
 {
     public class AdvancedRanges
     {
