+++ test/SemanticVersioning.Tests/LooseMode.cs
@@ -2,7 +2,7 @@ using System;
 using Xunit;
 using Xunit.Extensions;
 
-namespace SemVer.Tests
+namespace SemanticVersioning.Tests
 {
     public class LooseMode
     {
