+++ test/SemanticVersioning.Tests/PreReleaseComparisons.cs
@@ -1,7 +1,7 @@
 using Xunit;
 using Xunit.Extensions;
 
-namespace SemVer.Tests
+namespace SemanticVersioning.Tests
 {
     public class PreReleaseComparisons
     {
